photo-montage/requirements.txt
google-auth-oauthlib==1.2.1
google-api-python-client==2.164.0
requests==2.32.3
Pillow==11.1.0
numpy==2.2.4
python-dateutil==2.9.0.post0
